+++ components/app/context.ts
@@ -1,16 +1,20 @@
 import { inject, type InjectionKey } from 'vue'
 import type { AppConfig } from './types'
+import type { ModalFuncReturn } from '../modal/types'
 import { message } from '../message'
 import { notification } from '../notification'
 
-const noop = () => {}
+const noopModalFunc = (): ModalFuncReturn => ({
+  destroy: () => {},
+  update: () => {},
+})
 
 const defaultModal: AppConfig['modal'] = {
-  confirm: noop,
-  info: noop,
-  success: noop,
-  warning: noop,
-  error: noop,
+  confirm: noopModalFunc,
+  info: noopModalFunc,
+  success: noopModalFunc,
+  warning: noopModalFunc,
+  error: noopModalFunc,
 }
 
 export const APP_KEY: InjectionKey<AppConfig> = Symbol('App')