+++ components/app/index.ts
@@ -1,3 +1,3 @@
 export { App } from './App'
 export { useApp } from './context'
-export type { AppConfig } from './types'
+export type { AppConfig, AppProps } from './types'